fix(#430): address all 9 review findings — move semantics, stride API, path guard, shared maps
Review fixes:
  C1: set_depth_map takes DepthMap by value + std::move (avoid 1.2MB copy at 30Hz)
  C2: DepthMapUsedInFusion test forces past Tiers 1-3 and verifies ML depth magnitude
  W1: stride parameter on IDepthEstimator::estimate() for real camera row padding
  W2: Path traversal guard in load_model() rejects ".." in model paths
  W3: channels == 0 check (was channels < 1, semantically clearer for unsigned)
  W4: Extract class mapping functions to detector_class_maps.h (shared header)
  W5: Clarify Tier 3.5 comment — only fires when Tier 3 is geometrically impossible
  W6: Throttled failure logging + 50ms backpressure in depth_thread
  W7: Remove mutable from rng_/noise_dist_, document NOT thread-safe
