+++ common/hal/include/hal/idepth_estimator.h
@@ -30,13 +30,19 @@ public:
     virtual ~IDepthEstimator() = default;
 
     /// Estimate depth from a single RGB frame.
-    /// @param frame_data  Pointer to RGB pixel data (valid for this call only)
-    /// @param width       Frame width in pixels
-    /// @param height      Frame height in pixels
-    /// @param channels    Number of channels (3=RGB, 4=RGBA)
+    /// @param frame_data  Pointer to pixel data (valid for this call only).
+    /// @param width       Frame width in pixels (must be > 0)
+    /// @param height      Frame height in pixels (must be > 0)
+    /// @param channels    Number of channels (3=RGB, 4=RGBA; must be > 0)
+    /// @param stride      Row stride in bytes (0 = tightly packed, i.e. width * channels).
+    ///                    Some cameras pad rows for alignment; real backends must
+    ///                    respect stride when indexing pixels.
     /// @return DepthMap on success, error string on failure
+    /// @pre frame_data points to a buffer of at least height * stride bytes
+    ///      (or height * width * channels if stride == 0).
     [[nodiscard]] virtual drone::util::Result<DepthMap, std::string> estimate(
-        const uint8_t* frame_data, uint32_t width, uint32_t height, uint32_t channels) = 0;
+        const uint8_t* frame_data, uint32_t width, uint32_t height, uint32_t channels,
+        uint32_t stride = 0) = 0;
 
     /// Human-readable name (e.g. "SimulatedDepthEstimator", "DepthAnythingV2").
     [[nodiscard]] virtual std::string name() const = 0;
