+++ common/hal/include/hal/simulated_depth_estimator.h
@@ -15,6 +15,8 @@
 namespace drone::hal {
 
 /// Simulated depth estimator returning fixed depth with Gaussian noise.
+/// NOT thread-safe: rng_ is mutated on each estimate() call.
+/// Only call from a single thread (the depth thread in P2).
 /// Config keys:
 ///   perception.depth_estimator.default_depth_m  (default 10.0)
 ///   perception.depth_estimator.noise_std_m      (default 0.5)
@@ -34,17 +36,18 @@ public:
         , rng_(std::random_device{}())
         , noise_dist_(0.0f, std::max(0.0f, noise_std_m)) {}
 
-    [[nodiscard]] drone::util::Result<DepthMap, std::string> estimate(const uint8_t* frame_data,
-                                                                      uint32_t       width,
-                                                                      uint32_t       height,
-                                                                      uint32_t channels) override {
+    [[nodiscard]] drone::util::Result<DepthMap, std::string> estimate(
+        const uint8_t* frame_data, uint32_t width, uint32_t height, uint32_t channels,
+        uint32_t stride = 0) override {
+        // Simulated backend doesn't read pixel data — stride is unused.
+        (void)stride;
         if (frame_data == nullptr || width == 0 || height == 0) {
             return drone::util::Result<DepthMap, std::string>::err(
                 "SimulatedDepthEstimator: invalid frame (null data or zero dimensions)");
         }
-        if (channels < 1) {
+        if (channels == 0) {
             return drone::util::Result<DepthMap, std::string>::err(
-                "SimulatedDepthEstimator: invalid channel count");
+                "SimulatedDepthEstimator: invalid channel count (0)");
         }
 
         DepthMap map;
@@ -69,10 +72,10 @@ public:
     [[nodiscard]] std::string name() const override { return "SimulatedDepthEstimator"; }
 
 private:
-    float                                   default_depth_m_;
-    float                                   noise_std_m_;
-    mutable std::mt19937                    rng_;
-    mutable std::normal_distribution<float> noise_dist_;
+    float                           default_depth_m_;
+    float                           noise_std_m_;
+    std::mt19937                    rng_;
+    std::normal_distribution<float> noise_dist_;
 };
 
 }  // namespace drone::hal
