+++ process2_perception/include/perception/ifusion_engine.h
@@ -55,9 +55,9 @@ public:
     /// Default no-op — only UKFFusionEngine uses this.
     virtual void set_drone_pose(float /*north*/, float /*east*/, float /*up*/, float /*yaw*/) {}
 
-    /// Provide ML depth map for depth-enhanced fusion.
+    /// Provide ML depth map for depth-enhanced fusion (takes ownership via move).
     /// Default no-op — only UKFFusionEngine uses ML depth data.
-    virtual void set_depth_map(const drone::hal::DepthMap& /*depth_map*/) {}
+    virtual void set_depth_map(drone::hal::DepthMap /*depth_map*/) {}
 };
 
 /// Factory: create a fusion engine from a backend name.
